Move the repository global variables into RepoFactory
Allow RepoFactory to retain responsibility for creating the repository
objects, and allow the objects to be garbage collected when they are not
in use.
