+++ app/src/main/java/net/bible/service/download/RepoFactory.kt
@@ -16,12 +16,49 @@
  */
 package net.bible.service.download
 
+import androidx.annotation.VisibleForTesting
+import net.bible.service.common.CommonUtils
+import net.bible.service.sword.AcceptableBookTypeFilter
 import org.crosswire.jsword.book.Book
 
 /**
  * @author Martin Denham [mjdenham at gmail dot com]
  */
 class RepoFactory(val downloadManager: DownloadManager) {
+
+    private val andBibleRepo = Repository("AndBible", AcceptableBookTypeFilter())
+    private val andBibleExtraRepo = Repository("AndBible Extra", AcceptableBookTypeFilter())
+    private val andBibleBetaRepo = Repository(
+        "AndBible Beta",
+        object: AcceptableBookTypeFilter() {
+            override fun test(book: Book): Boolean = CommonUtils.isBeta
+        }
+    )
+
+    // see here for info ftp://ftp.xiphos.org/mods.d/
+    @VisibleForTesting
+    val crosswireRepo = Repository("CrossWire", AcceptableBookTypeFilter())
+    private val lockmanRepo = Repository("Lockman (CrossWire)", AcceptableBookTypeFilter())
+    private val wycliffeRepo = Repository("Wycliffe (CrossWire)", AcceptableBookTypeFilter())
+    private val crosswireBetaRepo = Repository(
+        "Crosswire Beta",
+        object : AcceptableBookTypeFilter() {
+            override fun test(book: Book): Boolean {
+                // just Calvin Commentaries for now to see how we go
+                //
+                // Cannot include Jasher, Jub, EEnochCharles because they are displayed as page per verse for some reason which looks awful.
+                if(CommonUtils.isBeta) return true
+                return super.test(book) &&
+                    book.initials == "CalvinCommentaries"
+            }
+        }
+    )
+
+    private val eBibleRepo = Repository("eBible", AcceptableBookTypeFilter())
+    private val stepRepo = Repository("STEP Bible (Tyndale)", AcceptableBookTypeFilter())
+    private val ibtRepo = Repository("IBT", AcceptableBookTypeFilter())
+
+
     private val defaultRepo = andBibleRepo
 
     // In priority order (if the same version of module is found in many, it will be picked up
